Move PoseWrt to a utility module
